# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hack Nation Genomic AST API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create Upload",
        "operationId": "create_upload_api_v1_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{upload_id}/content": {
      "put": {
        "tags": [
          "uploads"
        ],
        "summary": "Put Upload Content",
        "operationId": "put_upload_content_api_v1_uploads__upload_id__content_put",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_put_upload_content_api_v1_uploads__upload_id__content_put"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/uploads/{upload_id}": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "Get Upload",
        "operationId": "get_upload_api_v1_uploads__upload_id__get",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Upload Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Upload Api V1 Uploads  Upload Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Create Analysis",
        "operationId": "create_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAnalysisResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis Status",
        "operationId": "get_analysis_status_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatusResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/result": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis Result",
        "operationId": "get_analysis_result_api_v1_analyses__analysis_id__result_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}/events": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Stream Analysis Events",
        "operationId": "stream_analysis_events_api_v1_analyses__analysis_id__events_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisResult": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "uploading",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "sample": {
            "$ref": "#/components/schemas/SampleMetadata"
          },
          "qc": {
            "$ref": "#/components/schemas/QCReport"
          },
          "susceptibility": {
            "$ref": "#/components/schemas/SusceptibilityCall"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/VariantEvidence"
            },
            "type": "array",
            "title": "Variants"
          },
          "shap_features": {
            "items": {
              "$ref": "#/components/schemas/ShapFeature"
            },
            "type": "array",
            "title": "Shap Features"
          },
          "interpretation": {
            "$ref": "#/components/schemas/ClinicalInterpretation"
          },
          "pipeline_versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Pipeline Versions"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "status",
          "sample",
          "qc",
          "susceptibility",
          "variants",
          "shap_features",
          "interpretation",
          "pipeline_versions",
          "completed_at"
        ],
        "title": "AnalysisResult"
      },
      "AnalysisStatusResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "uploading",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0.0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AnalysisStatusResponse"
      },
...
      "Body_put_upload_content_api_v1_uploads__upload_id__content_put": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_put_upload_content_api_v1_uploads__upload_id__content_put"
      },
      "ClinicalInterpretation": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "key_drivers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Drivers"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "alternative_drugs": {
            "items": {
              "$ref": "#/components/schemas/AlternativeDrug"
            },
            "type": "array",
            "title": "Alternative Drugs"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Research use only. Not a clinical diagnostic. Confirm with phenotypic AST before treatment decisions."
          }
        },
        "type": "object",
        "required": [
          "summary",
          "key_drivers",
          "limitations",
          "alternative_drugs"
        ],
        "title": "ClinicalInterpretation"
      },
      "CreateAnalysisRequest": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "metadata": {
            "$ref": "#/components/schemas/SampleMetadata"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "object_key",
          "metadata"
        ],
        "title": "CreateAnalysisRequest"
      },
      "CreateAnalysisResponse": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "uploading",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "status",
          "created_at"
        ],
        "title": "CreateAnalysisResponse"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QCReport": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "file_format": {
            "type": "string",
            "enum": [
              "fasta",
              "fastq"
            ],
            "title": "File Format"
          },
          "total_bases": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bases"
          },
          "n50": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N50"
          },
          "contig_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contig Count"
          },
          "gc_content": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gc Content"
          },
          "species_call": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species Call"
          },
          "species_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species Confidence"
          },
          "contamination_flag": {
            "type": "boolean",
            "title": "Contamination Flag",
            "default": false
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "file_format"
        ],
        "title": "QCReport"
      },
      "SampleMetadata": {
        "properties": {
          "sample_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Sample Name"
          },
          "organism": {
            "type": "string",
            "title": "Organism",
            "default": "Escherichia coli"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          },
          "read_type": {
            "type": "string",
            "enum": [
              "short",
              "long",
              "assembly"
            ],
            "title": "Read Type",
            "default": "assembly"
          },
          "file_format": {
            "type": "string",
            "enum": [
              "fasta",
              "fastq"
            ],
            "title": "File Format"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "sample_name",
          "file_format"
        ],
        "title": "SampleMetadata"
      },
      "ShapFeature": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "shap_value": {
            "type": "number",
            "title": "Shap Value"
          },
          "direction": {
            "type": "string",
            "enum": [
              "resistant",
              "susceptible",
              "neutral"
            ],
            "title": "Direction"
          },
          "rank": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Rank"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "shap_value",
          "direction",
          "rank"
        ],
        "title": "ShapFeature"
      },
      "SusceptibilityCall": {
        "properties": {
          "drug": {
            "type": "string",
            "title": "Drug",
            "default": "ciprofloxacin"
          },
          "label": {
            "type": "string",
            "enum": [
              "R",
              "S",
              "I",
              "ATU",
              "unknown"
            ],
            "title": "Label"
          },
          "probability_resistant": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Probability Resistant"
          },
          "calibrated_probability": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibrated Probability"
          },
          "source": {
            "type": "string",
            "enum": [
              "ml",
              "amrfinderplus",
              "resfinder",
              "pointfinder",
              "reconciled"
            ],
            "title": "Source"
          },
          "breakpoint_standard": {
            "type": "string",
            "title": "Breakpoint Standard",
            "default": "EUCAST v16.1"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "label",
          "probability_resistant",
          "source",
          "confidence"
        ],
        "title": "SusceptibilityCall"
      },
      "UploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "default": "application/octet-stream"
          },
          "size_bytes": {
            "type": "integer",
            "maximum": 5000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes"
          },
          "metadata": {
            "$ref": "#/components/schemas/SampleMetadata"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "size_bytes",
          "metadata"
        ],
        "title": "UploadRequest"
      },
      "UploadResponse": {
        "properties": {
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds",
            "default": 3600
          }
        },
        "type": "object",
        "required": [
          "upload_id",
          "upload_url",
          "object_key"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VariantEvidence": {
        "properties": {
          "gene": {
            "type": "string",
            "title": "Gene"
          },
          "mutation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mutation"
          },
          "identity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identity"
          },
          "coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage"
          },
          "source": {
            "type": "string",
            "enum": [
              "ml",
              "amrfinderplus",
              "resfinder",
              "pointfinder",
              "reconciled"
            ],
            "title": "Source"
          },
          "associated_phenotype": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "R",
                  "S",
                  "I",
                  "ATU",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Associated Phenotype"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "gene",
          "source"
        ],
        "title": "VariantEvidence"
      }
    }
  }
}